Improved Database

+++ Database/Entities/Customer.cs
@@ -7,7 +7,7 @@ namespace Database.Entities
         public int Id { get; set; }
         public string Lastname { get; set; }
         public string Firstname { get; set; }
-        public int Phonenumber { get; set; }
+        public long Phonenumber { get; set; }
         public string Email { get; set; }
         public List<Order> Orders { get; set; }
     }
